Add Mean/Variance/Median for InverseGammaDistribution
Mean = Piecewise[{{b/(-1 + a), a > 1}}, Indeterminate],
Variance = Piecewise[{{b^2/((-2 + a)*(-1 + a)^2), a > 2}}, Indeterminate],
and Median = b/InverseGammaRegularized[a, 1/2] for the two-argument form,
matching wolframscript. Concrete shape parameters collapse the Piecewise
into the appropriate branch (e.g. a = 3 gives Mean = 1, Variance = 1)
or fall through to Indeterminate when the moment doesn't exist.

Also drops an unused-parens warning introduced by the previous
Hypergeometric2F1Regularized commit.

+++ src/functions/math_ast/hypergeometric.rs
@@ -787,7 +787,7 @@ fn hypergeometric_2f1_regularized_non_positive_c(
   if !(0..=64).contains(&m) {
     return Ok(None);
   }
-  let shift = (m + 1) as i128;
+  let shift = m + 1;
 
   // Pochhammer (a)_{m+1} as a Times product of (a, a+1, ..., a+m).
   let pochhammer = |x: &Expr| -> Expr {
